Always pass --32 to GNU AS backend unless -m64 was specified
At present, gcc calls GNU as with --32 iff gcc was called with -m32, and it
calls GNU as with --64 iff gcc itself was called with -m64.

This works fine assuming that GNU as works in 32-bit mode by default but not
otherwise.

This patch causes GNU as to be called with --64 if -m64 was passed to gcc
and to be called with --32 if -m64 was NOT passed to gcc.

Since the illumos gcc44 build outputs 32-bit by default, this causes the
backend assembler to be called with the correct options.